+++ packages/did-manager/src/index.ts
@@ -1,9 +1,11 @@
 import * as Interfaces from './interfaces';
 import * as MorpheusTransaction from './morpheus-transaction';
+import * as MorpheusNode from './node';
 import { TimeSeries } from './time-series';
 
 export {
   Interfaces,
   MorpheusTransaction,
+  MorpheusNode,
   TimeSeries,
 };