+++ packages/did-manager/src/node/index.ts
@@ -1,0 +1,4 @@
+export * from './block-handler';
+export * from './did-operations';
+export * from './layer2-api';
+export * from './transaction-handler';