+++ packages/did-manager/src/node/transaction-handler.ts
@@ -3,19 +3,19 @@ import { Database, State, TransactionPool } from '@arkecosystem/core-interfaces'
 import { Handlers } from '@arkecosystem/core-transactions';
 import { Interfaces as CryptoIf, Transactions } from '@arkecosystem/crypto';
 
-import { Interfaces } from '@internet-of-people/did-manager';
-import { Layer1, Types, Utils } from '@internet-of-people/sdk';
-
 import {
-  COMPONENT_NAME as READER_FACTORY_COMPONENT,
+  READER_FACTORY_COMPONENT_NAME,
   ITransactionReader,
   TransactionReaderFactory,
-} from './transaction-reader-factory';
+} from '@internet-of-people/hydra-plugin-core';
+
+import { Layer1, Types, Utils } from '@internet-of-people/sdk';
+import { IMorpheusStateHandler, MORPHEUS_STATE_HANDLER_COMPONENT_NAME } from '../interfaces/morpheus';
 
 /**
  * Handles Morpheus custom transactions on Layer 1 (IWalletManager)
  */
-export class MorpheusTransactionHandler extends Handlers.TransactionHandler {
+export class TransactionHandler extends Handlers.TransactionHandler {
   public getConstructor(): typeof Transactions.Transaction {
     return Layer1.MorpheusTransaction;
   }
@@ -34,10 +34,10 @@ export class MorpheusTransactionHandler extends Handlers.TransactionHandler {
 
     // Note: here we assume that when a block is reverted, the fact of the revert is NOT stored in the database.
     // This means, when a node is bootstrapped from scratch, it will not see any reverted blocks.
-    const readerFactory: TransactionReaderFactory = app.resolve(READER_FACTORY_COMPONENT);
+    const readerFactory: TransactionReaderFactory = app.resolve(READER_FACTORY_COMPONENT_NAME);
     const reader: ITransactionReader = await readerFactory(connection, this.getConstructor());
-    const stateHandler: Interfaces.IMorpheusStateHandler = app.resolve(
-      Interfaces.MORPHEUS_STATE_HANDLER_COMPONENT_NAME,
+    const stateHandler: IMorpheusStateHandler = app.resolve(
+      MORPHEUS_STATE_HANDLER_COMPONENT_NAME,
     );
 
     while (reader.hasNext()) {