+++ packages/hydra-plugin-core/src/index.ts
@@ -1,0 +1,2 @@
+export * from './transaction-reader-factory';
+export * from './interfaces';