+++ packages/hydra-plugin/src/main.ts
@@ -1,32 +1,26 @@
 import { Container, Database } from '@arkecosystem/core-interfaces';
 import { Handlers } from '@arkecosystem/core-transactions';
-import Optional from 'optional-js';
 import { asValue } from 'awilix';
 import { Server as HapiServer } from '@hapi/hapi';
+import Optional from 'optional-js';
 
-import { Interfaces, MorpheusTransaction } from '@internet-of-people/did-manager';
+import { Interfaces, MorpheusNode, MorpheusTransaction } from '@internet-of-people/did-manager';
 import { Interfaces as CryptoIf } from '@arkecosystem/crypto';
-import { Types, Utils } from '@internet-of-people/sdk';
-type TransactionId = Types.Sdk.TransactionId;
-
-import { MorpheusArkConnector } from './ark-connector';
-import { BlockEventSource } from './block-event-source';
-import { BlockHandler } from './block-handler';
-import { schedule } from './scheduler';
-import { Layer2API } from './layer2-api';
-import { MorpheusTransactionHandler } from './transaction-handler';
 import {
-  COMPONENT_NAME as READER_FACTORY_COMPONENT_NAME,
+  IInitializable,
+  READER_FACTORY_COMPONENT_NAME,
   transactionReaderFactory,
-} from './transaction-reader-factory';
+} from '@internet-of-people/hydra-plugin-core';
+import { Types, Utils } from '@internet-of-people/sdk';
+
+import { ArkConnector } from './ark-connector';
+import { BlockEventSource } from './block-event-source';
+import { schedule } from './scheduler';
 
 const { MorpheusStateHandler: { MorpheusStateHandler } } = MorpheusTransaction;
+type TransactionId = Types.Sdk.TransactionId;
 
-const PLUGIN_ALIAS = 'morpheus-hydra-plugin';
-
-export interface IInitializable {
-  init(): Promise<void>;
-}
+const PLUGIN_ALIAS = 'hydra-plugin';
 
 export class Composite implements IInitializable {
   private readonly log: Utils.IAppLog;
@@ -56,31 +50,29 @@ const attachHTTPApi = async(
   stateHandler: Interfaces.IMorpheusStateHandler,
   log: Utils.IAppLog,
 ): Promise<void> => {
-  log.info('Attaching Morpheus HTTP API...');
-
   const api = container.resolvePlugin('api');
 
   const database: Database.IDatabaseService = container.resolvePlugin('database');
   const transactionRepository: Database.ITransactionsBusinessRepository = database.transactionsBusinessRepository;
-  const txDb = {
+
+  const http: HapiServer = api.instance('http');
+
+  const morpheusLayer2API = new MorpheusNode.Layer2API(log, stateHandler, http, {
     getMorpheusTransaction: async(txId: TransactionId): Promise<Optional<Types.Layer1.IMorpheusAsset>> => {
       const txDetails: CryptoIf.ITransactionData = await transactionRepository.findById(txId);
       const morpheusTx = txDetails as Types.Layer1.IMorpheusData;
       return Optional.ofNullable(morpheusTx?.asset);
     },
-  };
-
-  const http: HapiServer = api.instance('http');
-  const layer2API = new Layer2API(log, stateHandler, http, txDb);
-  await layer2API.init();
-  log.info('HTTP API READY');
+  });
+  await morpheusLayer2API.init();
+  log.info('Morpheus HTTP API READY');
 };
 
 // TODO: separate register's content into a container, hence it can be tested
 const register = async(container: Container.IContainer): Promise<Composite> => {
   const log = new Utils.AppLog(container.resolvePlugin('logger'));
 
-  log.info('Starting up Morpheus');
+  log.info('Starting up Hydra Plugin...');
   const eventEmitter: NodeJS.EventEmitter = container.resolvePlugin('event-emitter');
   const blockEventSource = new BlockEventSource(
     log,
@@ -88,21 +80,24 @@ const register = async(container: Container.IContainer): Promise<Composite> => {
     schedule,
   );
 
-  // Cannot inject MorpheusTransactionHandler with the following values
-  // (the framework instantiates the handlers). We have to put its dependencies
-  // into the container, so the transaction handler can resolve them from there.
-  log.info('Creating State Handler');
-  const stateHandler = new MorpheusStateHandler(log, eventEmitter);
   container.register(READER_FACTORY_COMPONENT_NAME, asValue(transactionReaderFactory));
-  container.register(Interfaces.MORPHEUS_STATE_HANDLER_COMPONENT_NAME, asValue(stateHandler));
   container.register(Utils.LOGGER_COMPONENT_NAME, asValue(log));
 
-  const blockHandler = new BlockHandler(stateHandler, log);
+  // Cannot inject MorpheusStateHandler with the following values
+  // (the framework instantiates the handlers). We have to put its dependencies
+  // into the container, so the transaction handler can resolve them from there.
+  log.info('Creating Morpheus handlers...');
+  const stateHandler = new MorpheusStateHandler(log, eventEmitter);
+  container.register(Interfaces.MORPHEUS_STATE_HANDLER_COMPONENT_NAME, asValue(stateHandler));
 
-  const arkConnector = new MorpheusArkConnector(
+  const morpheusBlockHandler = new MorpheusNode.BlockHandler(stateHandler, log);
+
+  const arkConnector = new ArkConnector(
     eventEmitter,
     log,
-    blockHandler,
+    [
+      { subscriptionId: 'morpheus-block-listener', listener: morpheusBlockHandler },
+    ],
     blockEventSource,
   );
 
@@ -111,7 +106,7 @@ const register = async(container: Container.IContainer): Promise<Composite> => {
   await composite.init();
 
   log.info('Registering MorpheusTransactionHandler');
-  Handlers.Registry.registerTransactionHandler(MorpheusTransactionHandler);
+  Handlers.Registry.registerTransactionHandler(MorpheusNode.TransactionHandler);
 
   /* eslint @typescript-eslint/no-misused-promises: 0 */
   eventEmitter.on('wallet.api.started', async() => {